mpi: Change type of size from int to size_t
When measuring size, we should use size_t instead of int to avoid
warnings.

+++ src/mpi/pt2pt/bsendutil.c
@@ -78,8 +78,8 @@ static int initialized = 0;   /* keep track of the first call to any
 /* Forward references */
 static void MPIR_Bsend_retry_pending( void );
 static int MPIR_Bsend_check_active ( void );
-static MPII_Bsend_data_t *MPIR_Bsend_find_buffer( int );
-static void MPIR_Bsend_take_buffer( MPII_Bsend_data_t *, int );
+static MPII_Bsend_data_t *MPIR_Bsend_find_buffer( size_t );
+static void MPIR_Bsend_take_buffer( MPII_Bsend_data_t *, size_t);
 static int MPIR_Bsend_finalize( void * );
 static void MPIR_Bsend_free_segment( MPII_Bsend_data_t * );
 
@@ -527,7 +527,7 @@ static void MPIR_Bsend_retry_pending( void )
  * Find a slot in the avail buffer that can hold size bytes.  Does *not*
  * remove the slot from the avail buffer (see MPIR_Bsend_take_buffer) 
  */
-static MPII_Bsend_data_t *MPIR_Bsend_find_buffer( int size )
+static MPII_Bsend_data_t *MPIR_Bsend_find_buffer( size_t size )
 {
     MPII_Bsend_data_t *p = BsendBuffer.avail;
 
@@ -549,10 +549,10 @@ static MPII_Bsend_data_t *MPIR_Bsend_find_buffer( int size )
  * If there isn't enough left of p, remove the entire segment from
  * the avail list.
  */
-static void MPIR_Bsend_take_buffer( MPII_Bsend_data_t *p, int size  )
+static void MPIR_Bsend_take_buffer( MPII_Bsend_data_t *p, size_t size  )
 {
     MPII_Bsend_data_t *prev;
-    int         alloc_size;
+    size_t alloc_size;
 
     /* Compute the remaining size.  This must include any padding 
        that must be added to make the new block properly aligned */
@@ -563,11 +563,11 @@ static void MPIR_Bsend_take_buffer( MPII_Bsend_data_t *p, int size  )
        allocate for this buffer. */
 
     MPL_DBG_MSG_FMT(MPIR_DBG_BSEND,TYPICAL,(MPL_DBG_FDEST,
-                                    "Taking %d bytes from a block with %llu bytes\n",
+                                    "Taking %lu bytes from a block with %llu bytes\n",
                                     alloc_size, (unsigned long long) p->total_size ));
 
     /* Is there enough space left to create a new block? */
-    if (alloc_size + (int)BSENDDATA_HEADER_TRUE_SIZE + MIN_BUFFER_BLOCK <= p->size) {
+    if (alloc_size + BSENDDATA_HEADER_TRUE_SIZE + MIN_BUFFER_BLOCK <= p->size) {
 	/* Yes, the available space (p->size) is large enough to 
 	   carve out a new block */
 	MPII_Bsend_data_t *newp;
